move page to correct folder
