+++ frontend/src/pages/trade/MyTrades.tsx
